requirements.txt
Flask>=3.0,<4.0
reportlab>=4.2,<5.0
python-docx>=1.1,<2.0
gunicorn>=22.0,<24.0
